switch to cmus status_display_program for status
For py3status cmus module, rather than querying the status
with cmus-remote, use the built-in status_display_program
functionality to write the status to a text file on any
status change.
Then the py3status module displays the contents of that file
but converts the status indicator to a user-configurable symbol
and assigns the appropriate colors.